+++ arandomness/tests/test_agenerator.py
@@ -1,9 +1,9 @@
 #! /usr/bin/env python
 
-"""Test arandomness' autocorrect
+"""<description>
 
 Copyright:
-    test_autocorrect.py  test arandomness' autocorrect
+    <script name>  <description>
     Copyright (C) 2017  Alex Hyer
 
     This program is free software: you can redistribute it and/or modify
@@ -20,34 +20,14 @@ Copyright:
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
-from ..str import autocorrect
+from ..arandom import agenerator
 
 __author__ = 'Alex Hyer'
 __email__ = 'theonehyer@gmail.com'
 __license__ = 'GPLv3'
 __maintainer__ = 'Alex Hyer'
-__status__ = 'Production/Stable'
-__version__ = '1.0.1'
+__status__ = 'Planning'
+__version__ = '0.1.0a1'
 
-
-def test_autocorrect():
-    """Test arandomness' autocorrect"""
-
-    # Setup test data
-    words1 = ['bowtie2', 'bot']
-    words2 = ['apple', 'apocalyptic', 'off word']
-    words3 = ['computer', 'computational', 'compute']
-    words4 = ['hello', 'hi']
-
-    # Test "normal" word lists
-    assert autocorrect('bow', words1) == 'bowtie2'
-    assert autocorrect('apo', words2) == 'apocalyptic'
-
-    # Test when no match available
-    try:
-        autocorrect('bad', words3, delta=0.75) == ''
-    except AssertionError as error:
-        assert str(error) == 'No matches for "bad" found'
-
-    # Test exact word match
-    assert autocorrect('hello', words4) == 'hello'
+def test_agenerator():
+    assert agenerator() == 10