+++ arandomness/tests/test_max_substring.py
@@ -20,14 +20,14 @@ Copyright:
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 
-from arandomness.str import max_substring
+from ..str import max_substring
 
 __author__ = 'Alex Hyer'
 __email__ = 'theonehyer@gmail.com'
 __license__ = 'GPLv3'
 __maintainer__ = 'Alex Hyer'
 __status__ = 'Production/Stable'
-__version__ = '1.0.0'
+__version__ = '1.0.1'
 
 
 def test_max_substring():